src/requirements.txt
asgiref==3.2.10
Django==3.1.2
pytz==2020.1
sqlparse==0.4.0
psycopg2-binary==2.8.5
gunicorn==20.0.4
djangorestframework==3.12.1
django-cors-headers==3.5.0
django-rest-auth==0.9.5
